Conflicts:
	bin/R.txt
	bin/ZF_Agent_A_Pad.apk
	bin/classes.dex
	bin/classes/com/comdo/zf_agent_a_pad/activity/MainActivity.class
	bin/classes/com/comdo/zf_agent_a_pad/util/ImageCacheUtil$1.class
	bin/classes/com/comdo/zf_agent_a_pad/util/SetPopWindow.class
	bin/classes/com/comdo/zf_agent_a_pad/util/XListView.class
	bin/classes/com/comdo/zf_agent_a_pad/util/XListViewHeader.class
	bin/classes/com/example/zf_agent_a_pad/R$drawable.class
	bin/classes/com/example/zf_agent_a_pad/R$id.class
	bin/classes/com/example/zf_agent_a_pad/R$layout.class
	bin/classes/com/example/zf_agent_a_pad/R$menu.class
	bin/classes/com/example/zf_agent_a_pad/R$string.class
	bin/classes/com/example/zf_agent_a_pad/R$style.class
	bin/classes/com/example/zf_agent_a_pad/R$styleable.class
	bin/resources.ap_
	gen/com/example/zf_agent_a_pad/R.java
	project.properties
	res/layout/fragment_trade_flow_list.xml
	src/com/comdo/zf_agent_a_pad/trade/TradeAgentActivity.java
	src/com/comdo/zf_agent_a_pad/trade/TradeClientActivity.java
	src/com/comdo/zf_agent_a_pad/trade/TradeDetailActivity.java
	src/com/comdo/zf_agent_a_pad/trade/TradeFlowActivity.java
	src/com/comdo/zf_agent_a_pad/trade/TradeStatisticActivity.java

+++ src/com/comdo/zf_agent_a_pad/entity/StaffmanagerEntity.java
@@ -1,21 +1,21 @@
 package com.comdo.zf_agent_a_pad.entity;
 
 public class StaffmanagerEntity {
-	private long id;
+	private int id;
 	private String username;
 	private String name;
 	private String createdAt;
-	public StaffmanagerEntity(long id,String username,String name,String createdAt){
+	public StaffmanagerEntity(int id,String username,String name,String createdAt){
 		super();
 		this.id=id;
 		this.username=username;
 		this.name=name;
 		this.createdAt=createdAt;
 	}
-	public long getId() {
+	public int getId() {
 		return id;
 	}
-	public void setId(long id) {
+	public void setId(int id) {
 		this.id = id;
 	}
 	public String getUsername() {
